+++ tools/Beef.CodeGen.Core/Config/Entity/CodeGenConfig.cs
@@ -31,6 +31,7 @@ entities:
     [CategorySchema("EntityFramework", Title = "Provides the _Entity Framewotrk (EF) Data-layer_ configuration.")]
     [CategorySchema("Cosmos", Title = "Provides the _CosmosDB Data-layer_ configuration.")]
     [CategorySchema("OData", Title = "Provides the _OData Data-layer_ configuration.")]
+    [CategorySchema("HttpAgent", Title = "Provides the _HTTP Agent Data-layer_ configuration.")]
     [CategorySchema("gRPC", Title = "Provides the _gRPC_ configuration.")]
     [CategorySchema("Path", Title = "Provides the _Path (Directory)_ configuration for the generated artefacts.")]
     [CategorySchema("Namespace", Title = "Provides the _.NET Namespace_ configuration for the generated artefacts.")]
@@ -236,6 +237,14 @@ entities:
         [PropertySchema("OData", Title = "The .NET OData interface name used where `Operation.AutoImplement` is `OData`.", IsImportant = true,
             Description = "Defaults to `IOData`. This can be overridden within the `Entity`(s).")]
         public string? ODataName { get; set; }
+
+        /// <summary>
+        /// Gets or sets the default .NET HTTP Agent interface name used where `Operation.AutoImplement` is `HttpRest`.
+        /// </summary>
+        [JsonProperty("httpAgentName", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        [PropertySchema("HttpAgent", Title = "The default .NET HTTP Agent interface name used where `Operation.AutoImplement` is `HttpAgent`.", IsImportant = true,
+            Description = "Defaults to `IHttpAgent`. This can be overridden within the `Entity`(s).")]
+        public string? HttpAgentName { get; set; }
 
         /// <summary>
         /// Gets or sets the default Reference Data property Converter used by the generated Mapper(s) where not specifically defined.
@@ -632,6 +641,7 @@ entities:
             EntityFrameworkName = DefaultWhereNull(EntityFrameworkName, () => "IEfDb");
             CosmosName = DefaultWhereNull(CosmosName, () => "ICosmosDb");
             ODataName = DefaultWhereNull(ODataName, () => "IOData");
+            HttpAgentName = DefaultWhereNull(HttpAgentName, () => "IHttpAgent");
             JsonSerializer = DefaultWhereNull(JsonSerializer, () => "Newtonsoft");
             ETagJsonName = DefaultWhereNull(ETagJsonName, () => "etag");
             RefDataDefaultMapperConverter = DefaultWhereNull(RefDataDefaultMapperConverter, () => "ReferenceDataCodeConverter<T>");